Separarted client and server directories
