+++ .dockerignore
@@ -1,1 +1,45 @@
-Dockerfile
+.idea/
+client/node_modules/
+client/js/app.js
+client/js/vendor.js
+client/js/index.html
+server/cmd/*.exe
+server/cmd/logs/
+server/cmd/tmp/
+
+# Compiled source #
+###################
+*.class
+*.dll
+*.exe
+*.o
+*.so
+
+# Packages #
+############
+# it's better to unpack these files and commit the raw source
+# git has its own built in compression methods
+*.7z
+*.dmg
+*.gz
+*.iso
+*.jar
+*.rar
+*.tar
+*.zip
+
+# Logs and databases #
+######################
+*.log
+*.sql
+*.sqlite
+
+# OS generated files #
+######################
+.DS_Store
+.DS_Store?
+._*
+.Spotlight-V100
+.Trashes
+ehthumbs.db
+Thumbs.db
