+++ client/package.json
@@ -38,7 +38,7 @@
     "html-webpack-harddisk-plugin": "^0.2.0",
     "html-webpack-plugin": "^3.2.0",
     "mobx-react-devtools": "^5.0.1",
-    "node-sass": "^4.8.3",
+    "node-sass": "^4.11.0",
     "sass-loader": "^7.0.3",
     "style-loader": "^0.21.0",
     "ts-loader": "^4.1.0",
